# openapi/api.yaml
components:
  schemas:
...
    ExportAsHTMLRequest:
      properties:
        assetUrl:
          nullable: true
          type: string
        download:
          type: boolean
        files:
          items:
            type: string
          type: array
        includeCode:
          type: boolean
      required:
      - download
      - files
      - includeCode
      type: object
    ExportAsIPYNBRequest:
      properties:
        download:
          type: boolean
      required:
      - download
      type: object
    ExportAsMarkdownRequest:
      properties:
        download:
          type: boolean
      required:
      - download
      type: object
    ExportAsScriptRequest:
      properties:
        download:
          type: boolean
      required:
      - download
      type: object
...
    SuccessResponse:
      properties:
        success:
          type: boolean
      required:
      - success
      type: object
...
info:
  title: marimo API
  version: 0.10.2
openapi: 3.1.0
paths:
...
  /api/export/auto_export/html:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsHTMLRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: Export the notebook as HTML
        400:
          description: File must be saved before downloading
  /api/export/auto_export/ipynb:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsIPYNBRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: Export the notebook as IPYNB
        400:
          description: File must be saved before downloading
  /api/export/auto_export/markdown:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsMarkdownRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: Export the notebook as a markdown
        400:
          description: File must be saved before downloading
  /api/export/html:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsHTMLRequest'
      responses:
        200:
          content:
            text/html:
              schema:
                type: string
          description: Export the notebook as HTML
        400:
          description: File must be saved before downloading
  /api/export/markdown:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsMarkdownRequest'
      responses:
        200:
          content:
            text/plain:
              schema:
                type: string
          description: Export the notebook as a markdown
        400:
          description: File must be saved before downloading
  /api/export/script:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportAsScriptRequest'
      responses:
        200:
          content:
            text/plain:
              schema:
                type: string
          description: Export the notebook as a script
        400:
          description: File must be saved before downloading
...